LayoutTests/fast/dom/shadow/shadow-disable-expected.txt

Tests to ensure that shadow element cannot be created in elements having dynamically created shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


DIV
PASS new WebKitShadowRoot(element) is not null
SPAN
PASS new WebKitShadowRoot(element) is not null
A
PASS new WebKitShadowRoot(element) is not null
SECTION
PASS new WebKitShadowRoot(element) is not null
AUDIO
PASS new WebKitShadowRoot(element) is not null
VIDEO
PASS new WebKitShadowRoot(element) is not null
SELECT
PASS new WebKitShadowRoot(element) is not null
INPUT
PASS new WebKitShadowRoot(element) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
TEXTAREA
PASS new WebKitShadowRoot(element) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
tref
PASS new WebKitShadowRoot(element) threw exception Error: HIERARCHY_REQUEST_ERR: DOM Exception 3.
PASS successfullyParsed is true

TEST COMPLETE
